feat: Hide `Tooltip` when pressing `Escape`
